+++ tests/integration/discovery/test_discovery_pandas_only_IT.py
@@ -5,44 +5,9 @@ from neo4j_runway.discovery import Discovery
 from neo4j_runway.utils.data import Table, TableCollection
 from neo4j_runway.warnings import ExperimentalFeatureWarning
 
-data_dict = {
-    "a.csv": {"a": "numbers", "b": "more numbers"},
-    "b.csv": {"c": "many more numbers", "d": "lots of numbers"},
-    "c.csv": {"e": "letters", "f": "chars"},
-}
-t1 = Table(
-    name="a.csv",
-    file_path="./a.csv",
-    dataframe=pd.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6]}),
-    data_dictionary=data_dict["a.csv"],
-    use_cases=["test discovery"],
-)
-t2 = Table(
-    name="b.csv",
-    file_path="./b.csv",
-    dataframe=pd.DataFrame({"c": [7, 8, 9], "d": [10, 11, 12]}),
-    data_dictionary=data_dict["b.csv"],
-    use_cases=["test discovery"],
-)
-t3 = Table(
-    name="c.csv",
-    file_path="./c.csv",
-    dataframe=pd.DataFrame({"e": ["a", "b", "c"], "f": ["d", "e", "f"]}),
-    data_dictionary=data_dict["c.csv"],
-    use_cases=["test discovery"],
-)
-table_collection = TableCollection(
-    data_directory="./",
-    data_dictionary=data_dict,
-    tables=[t1, t2, t3],
-    general_description="contain data for testing discovery",
-    use_cases=["test discovery"],
-)
-
 
 def test_single_dataframe_run_pandas_only() -> None:
-    with pytest.warns(UserWarning):
-        d = Discovery(data=pd.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6]}))
+    d = Discovery(data=pd.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6]}))
 
     d.run(pandas_only=True)
 
@@ -51,8 +16,8 @@ def test_single_dataframe_run_pandas_only() -> None:
     assert d.data.tables[0].discovery is not None
 
 
-def test_single_table_run_pandas_only() -> None:
-    d = Discovery(data=t1)
+def test_single_table_run_pandas_only(table_1: Table) -> None:
+    d = Discovery(data=table_1)
 
     d.run(pandas_only=True)
 
@@ -61,7 +26,7 @@ def test_single_table_run_pandas_only() -> None:
     assert d.data.tables[0].discovery is not None
 
 
-def test_multi_file_run_pandas_only() -> None:
+def test_multi_file_run_pandas_only(table_collection: TableCollection) -> None:
     with pytest.warns(ExperimentalFeatureWarning):
         d = Discovery(data=table_collection)