+++ tests/unit/modeler/test_data_modeler_llm_calls.py
@@ -3,19 +3,15 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from neo4j_runway.modeler import GraphDataModeler
+from neo4j_runway.utils.data.data_dictionary.data_dictionary import DataDictionary
 from neo4j_runway.warnings import ExperimentalFeatureWarning
-
-data_dictionary = {
-    "a.csv": {"a": "test", "b": "test2"},
-    "b.csv": {"c": "test3"},
-}
 
 
 @patch(
     "neo4j_runway.llm.openai.data_modeling.OpenAIDataModelingLLM._get_data_model_response"
 )
 def test_create_initial_model(
-    mock_llm: MagicMock,
+    mock_llm: MagicMock, data_dictionary: DataDictionary
 ) -> None:
     with pytest.warns((UserWarning, ExperimentalFeatureWarning)):
         gdm = GraphDataModeler(llm=mock_llm, data_dictionary=data_dictionary)
@@ -26,8 +22,7 @@ def test_create_initial_model(
 
 
 def test_iterate_model(
-    mock_llm: MagicMock,
-    mock_data_model: MagicMock,
+    mock_llm: MagicMock, mock_data_model: MagicMock, data_dictionary: DataDictionary
 ) -> None:
     with pytest.warns((UserWarning, ExperimentalFeatureWarning)):
         gdm = GraphDataModeler(llm=mock_llm, data_dictionary=data_dictionary)