+++ tests/unit/models/core/test_data_model_validation.py
@@ -19,19 +19,10 @@ def test_multi_file_different_source_relationship_valid_target_node(
 
 
 def test_multi_file_different_source_relationship_invalid(
-    data_model_data: Dict[str, Any], data_model_context: Dict[str, Any]
+    data_model_data: Dict[str, Any], data_model_bad_context: Dict[str, Any]
 ) -> None:
-    data_dictionary_bad = {
-        "a.csv": {"id": "unique id for a. Has alias a_id", "b": "feature b"},
-        "b.csv": {"id": "unique id for b", "c": "feature c"},
-    }
-    valid_columns_bad = {k: list(v.keys()) for k, v in data_dictionary_bad.items()}
-
-    data_model_context["data_dictionary"] = data_dictionary_bad
-    data_model_context["valid_columns"] = valid_columns_bad
-
     with pytest.raises(ValidationError) as e:
-        DataModel.model_validate(data_model_data, context=data_model_context)
+        DataModel.model_validate(data_model_data, context=data_model_bad_context)
 
     assert (
         "relationship_source_file_missing_source_node_unique_property_alias_error"