+++ CHANGELOG.md
@@ -6,7 +6,12 @@
 
 ### Changed
 
+* Deprecating `user_input` args and `UserInput` object. The resposibilities of these are handled by `TableCollection` and `DataDictionary`
+* Removed integration tests that required connection to LLM endpoints
+
 ### Added
+
+* Data dictionaries are now held in a `DataDictionary` object allowing for advanced validation and better integrations throughout the code base
 
 ## 0.14.0